requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
SQLAlchemy==2.0.18
argon2-cffi==21.3.0
python-multipart==0.0.6
python-jose==3.3.0